Tooltip changes to fit hud
The matchmaking dashboard right click does not act as a tooltip and instead acts like a combo box
It does not have a .res file to modify so instead I had to make changes to the font file directly and noted the colors in basesettings (Menu).  The border is controlled by NotificationDefault.
Did not make changes to the bg/fg/border colors